# packages/app-contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TreeTime API",
    "description": "Maximum-likelihood phylodynamic inference",
    "contact": {
      "name": "NeherLab",
      "url": "https://github.com/neherlab/treetime"
    },
    "license": {
      "name": "MIT"
    },
    "version": "1.0.0"
  },
  "paths": {
    "/api/check-config": {
      "post": {
        "operationId": "configCheck",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckConfigRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The configuration with every default filled in, or the problems found in it",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckConfigResponse"
                }
              }
            }
          }
        },
        "x-bridge-type": "request"
      }
    },
    "/api/check-inputs": {
      "post": {
        "operationId": "inputsCheck",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckInputsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Facts about the input files, read with the readers the commands use",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InputFacts"
                }
              }
            }
          }
        },
        "x-bridge-type": "request"
      }
    },
...
    "/api/runs": {
      "get": {
        "operationId": "runsList",
        "responses": {
          "200": {
            "description": "Runs, newest first, and the number of runs computing now",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunList"
                }
              }
            }
          }
        },
        "x-bridge-type": "query"
      },
      "post": {
        "operationId": "runsCreate",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRunRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The created run; it starts at once unless `defer_start` is set",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunRecord"
                }
              }
            }
          }
        },
        "x-bridge-type": "request"
      }
    },
    "/api/runs/{id}": {
      "get": {
        "operationId": "runsGet",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of the run",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The record of the run",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunRecord"
                }
              }
            }
          }
        },
        "x-bridge-type": "query"
      },
      "delete": {
        "operationId": "runsDelete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of the run",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "The run moved to the trash; restore undoes it"
          }
        },
        "x-bridge-type": "request"
      },
      "patch": {
        "operationId": "runsUpdate",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of the run",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRunRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The run with its new title or pinned state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunSummary"
                }
              }
            }
          }
        },
        "x-bridge-type": "request"
      }
    },
    "/api/runs/{id}/archive": {
      "get": {
        "operationId": "runsArchive",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of the run",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Zip archive of the run's `out/` folder",
            "content": {
              "application/zip": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        },
        "x-bridge-type": "file"
      }
    },
    "/api/runs/{id}/cancel": {
      "post": {
        "operationId": "runsCancel",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of the run",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Whether cancellation was requested; the run ends with a `cancelled` terminal event",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CancelRunResponse"
                }
              }
            }
          }
        },
        "x-bridge-type": "request"
      }
    },
    "/api/runs/{id}/events": {
      "get": {
        "operationId": "runsEvents",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of the run",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "in": "query",
            "description": "Sequence number of the first event to send; the `Last-Event-ID` header of a reconnect resumes after that event",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Stream of the run's events from `from`: `started`, `progress`, `log` and `iteration`, then one `terminal`",
            "content": {
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/RunEvent"
                }
              }
            }
          }
        },
        "x-bridge-type": "stream"
      }
    },
    "/api/runs/{id}/file": {
      "get": {
        "operationId": "runsFile",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of the run",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "path",
            "in": "query",
            "description": "Path of the file relative to the run's `out/` folder",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Contents of the file",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        },
        "x-bridge-type": "file"
      }
    },
    "/api/runs/{id}/files": {
      "get": {
        "operationId": "runsFiles",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of the run",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Files in the run's `out/` folder, with their sizes and kinds",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RunFile"
                  }
                }
              }
            }
          }
        },
        "x-bridge-type": "query"
      }
    },
    "/api/runs/{id}/inputs/{name}": {
      "put": {
        "operationId": "runsUploadInput",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of a run that has not started",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "path",
            "description": "File name inside the run's `inputs/` folder",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The stored file and the path to use for it in the run's configuration",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadedInput"
                }
              }
            }
          },
          "413": {
            "description": "The inputs of the run exceed the upload limit of the server",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "x-bridge-type": "upload",
        "requestBody": {
          "required": true,
          "content": {
            "application/octet-stream": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            }
          }
        }
      }
    },
    "/api/runs/{id}/purge": {
      "post": {
        "operationId": "runsPurge",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of a deleted run",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "The deleted run is removed for good"
          }
        },
        "x-bridge-type": "request"
      }
    },
    "/api/runs/{id}/restore": {
      "post": {
        "operationId": "runsRestore",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of a deleted run",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The run, back from the trash",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunSummary"
                }
              }
            }
          }
        },
        "x-bridge-type": "request"
      }
    },
    "/api/runs/{id}/start": {
      "post": {
        "operationId": "runsStart",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of the run",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartRunRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The started run",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunRecord"
                }
              }
            }
          }
        },
        "x-bridge-type": "request"
      }
    },
...
  },
  "components": {
    "schemas": {
      "CancelRunResponse": {
        "type": "object",
        "required": [
          "cancelled"
        ],
        "properties": {
          "cancelled": {
            "type": "boolean"
          }
        }
      },
...
      "CheckConfigRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "command": {
            "$ref": "#/components/schemas/AppCommand",
            "description": "Command the configuration is for."
          },
          "text": {
            "type": "string",
            "description": "Configuration as YAML or JSON text."
          }
        },
        "required": [
          "command",
          "text"
        ],
        "description": "Request to check a configuration."
      },
...
      "RunList": {
        "type": "object",
        "properties": {
          "runs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RunSummary"
            },
            "description": "Runs, newest first."
          },
          "active_runs": {
            "type": "integer",
            "format": "uint",
            "minimum": 0,
            "description": "Number of runs whose computation is running now. Every run shares one pool of processing threads."
          }
        },
        "required": [
          "runs",
          "active_runs"
        ],
        "description": "Runs, newest first, and the number of runs computing now."
      },
...
    }
  }
}